libutil/url: fix git+file:./ parse error
Previously, the "file:./" prefix was not correctly recognized in
fixGitURL; instead, it was mistaken as a file path, which resulted in a
parsed url of the form "file://file:./".

This commit fixes the issue by properly detecting the "file:" prefix.
Note, however, that unlike "file://", the "file:./" URI is _not_
standardized, but is widely used to referred to relative file paths.
In particular, the "git+file:./" did work for nix<=2.18, and was broken
in nix 2.19.0.

A test is added to "tests/functional/fetchGit.sh" to prevent future
breakage.